+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ci: algo endpoint now returns 404 instead of 400 for non-uuid keys in retrieve calls (#207)
 - ci: datamanager endpoint now returns 404 instead of 400 for non-uuid keys in retrieve calls (#209)
 - ci: tuple endpoints now returns 404 instead of 400 for non-uuid keys in retrieve calls (#216)
+- ci: compute plan endpoint now returns 404 instead of 400 for non-uuid keys in retrieve calls (#219)
 
 ## [0.13.0] - 2022-03-01